Day16 - cl compiler options & make build easier & keyboard input fix

+++ handmade.cpp
@@ -104,8 +104,8 @@ static void render_weird_gradient(GameOffscreenBuffer *buffer, int x_offset, int
     uint32_t *arr = (uint32_t *)buffer->memory;
     for (int y = 0; y < buffer->height; y++) {
         for (int x = 0; x < buffer->width; x++) {
-            uint8_t blue = x + x_offset;
-            uint8_t green = y + y_offset;
+            uint8_t blue = (uint8_t)(x + x_offset);
+            uint8_t green = (uint8_t)(y + y_offset);
             *arr++ = ((green << 8) | blue);
         }
     }
